support cloning instead of fork and clone operation
The original idea was for people to fork and clone and then create their own branch for their changes, however, it is easier to just clone this project and make changes. No need to maintain a remote upstream unless they want to contribute to this project.

Changes to be committed:

	modified:   config.sh
	modified:   init.sh
	modified:   setup.sh